Has simplified getting a step description
