+++ stepik-union/src/main/java/org/stepik/core/courseFormat/stepHelpers/ChoiceQuizNodeHelper.java
@@ -1,8 +1,8 @@
 package org.stepik.core.courseFormat.stepHelpers;
 
 import com.intellij.openapi.util.Pair;
-import org.stepik.core.courseFormat.StepNode;
 import org.jetbrains.annotations.NotNull;
+import org.stepik.core.courseFormat.StepNode;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -10,13 +10,13 @@ import java.util.List;
 /**
  * @author meanmail
  */
-public class ChoiceStepNodeHelper extends StepHelper {
+public class ChoiceQuizNodeHelper extends QuizHelper {
     private boolean isMultipleChoice;
     private List<Pair<String, Boolean>> stepOptions;
     private String[] options;
     private List<Boolean> choices;
 
-    public ChoiceStepNodeHelper(@NotNull StepNode stepNode) {
+    public ChoiceQuizNodeHelper(@NotNull StepNode stepNode) {
         super(stepNode);
     }
