+++ stepik-union/src/main/java/org/stepik/core/courseFormat/stepHelpers/StringQuizNodeHelper.java
@@ -1,17 +1,17 @@
 package org.stepik.core.courseFormat.stepHelpers;
 
-import org.stepik.core.courseFormat.StepNode;
 import org.apache.commons.lang.StringEscapeUtils;
 import org.jetbrains.annotations.NotNull;
+import org.stepik.core.courseFormat.StepNode;
 
 /**
  * @author meanmail
  */
-public class StringStepNodeHelper extends StepHelper {
+public class StringQuizNodeHelper extends QuizHelper {
     private boolean isTextDisabled;
     private String text;
 
-    public StringStepNodeHelper(@NotNull StepNode stepNode) {
+    public StringQuizNodeHelper(@NotNull StepNode stepNode) {
         super(stepNode);
     }
