+++ stepik-union/src/main/java/org/stepik/core/courseFormat/stepHelpers/TextHelper.java
@@ -1,0 +1,18 @@
+package org.stepik.core.courseFormat.stepHelpers;
+
+import org.jetbrains.annotations.NotNull;
+import org.stepik.core.courseFormat.StepNode;
+
+/**
+ * @author meanmail
+ */
+public class TextHelper extends StepHelper {
+    public TextHelper(@NotNull StepNode stepNode) {
+        super(stepNode);
+    }
+
+    @NotNull
+    public String getLinkTitle() {
+        return "Read this step on Stepik";
+    }
+}
