+++ stepik-union/src/main/resources/templates/quiz/base.ftl
@@ -1,12 +1,14 @@
-<#-- @ftlvariable name="stepNode" type="org.stepik.core.courseFormat.stepHelpers.StepHelper" -->
+<#-- @ftlvariable name="stepNode" type="org.stepik.core.courseFormat.stepHelpers.QuizHelper" -->
 <#-- @ftlvariable name="text" type="java.lang.String" -->
-<#macro quiz_content>
-${text}<br>
+<#include "base_step.ftl">
 
+<@step_content>
+
+<#macro quiz_content>
 <div>
     <#assign status = stepNode.getStatus()/>
-    <#assign isHasSubmissionsRestrictions = stepNode.getStepNode().getData().isHasSubmissionsRestrictions() />
-    <#assign maxSubmissionsCount = stepNode.getStepNode().getData().getMaxSubmissionsCount() />
+    <#assign isHasSubmissionsRestrictions = stepNode.isHasSubmissionsRestrictions() />
+    <#assign maxSubmissionsCount = stepNode.getMaxSubmissionsCount() />
     <#assign submissionsCount = stepNode.getSubmissionsCount() />
     <#assign locked = isHasSubmissionsRestrictions && (submissionsCount >= maxSubmissionsCount) />
 
@@ -75,3 +77,4 @@ ${text}<br>
 
 </div>
 </#macro>
+</@step_content>
