+++ stepik-union/src/main/resources/templates/quiz/choice.ftl
@@ -1,5 +1,5 @@
 <#-- @ftlvariable name="disabled" type="java.lang.String" -->
-<#-- @ftlvariable name="stepNode" type="org.stepik.core.courseFormat.stepHelpers.ChoiceStepNodeHelper" -->
+<#-- @ftlvariable name="stepNode" type="org.stepik.core.courseFormat.stepHelpers.ChoiceQuizNodeHelper" -->
 
 <#include "base.ftl">
