+++ stepik-union/src/main/resources/templates/quiz/code.ftl
@@ -1,0 +1,11 @@
+<#-- @ftlvariable name="sample" type="org.stepik.api.objects.steps.Sample" -->
+
+<#include "base_step.ftl">
+
+<@step_content>
+
+${stepNode.getSamples()}
+
+<p><b>Limits:</b> ${stepNode.getTimeLimit()} s; ${stepNode.getMemoryLimit()} Mib</p>
+
+</@step_content>
