+++ stepik-union/src/main/resources/templates/quiz/unknown.ftl
@@ -1,0 +1,5 @@
+<#include "base_step.ftl">
+
+<@step_content>
+
+</@step_content>
